Remove usage of atomic from oscap-docker
The last release of atomic was 5 years ago.
The project is officially deprecated.
No linux distributions still package atomic.
