+++ utils/oscap-docker.in
@@ -21,8 +21,7 @@
 ''' oscap docker command '''
 
 import argparse
-from oscap_docker_python.oscap_docker_util import OscapAtomicScan, \
-    OscapDockerScan, isAtomicLoaded
+from oscap_docker_python.oscap_docker_util import OscapDockerScan
 
 import docker
 import traceback
@@ -47,9 +46,6 @@ if __name__ == '__main__':
                                      more about OSCAP-ARGUMENTS')
     parser.add_argument('--oscap', dest='oscap_binary', default='',
                         help='Set the oscap binary to use')
-
-    parser.add_argument('--disable-atomic', dest='noatomic', action='store_true',
-                        help="Force to use native docker API instead of atomic")
     subparser = parser.add_subparsers(help="commands")
 
     # Scan CVEs in image
@@ -94,28 +90,14 @@ if __name__ == '__main__':
         sys.exit(1)
 
     try:
-        if isAtomicLoaded() and not args.noatomic:
-            print("Using Atomic API")
-            OS = OscapAtomicScan(oscap_binary=args.oscap_binary)
-            if args.action == "scan":
-                rc = OscapAtomicScan.scan(OS, args.scan_target, leftover_args)
-            elif args.action == "scan_cve":
-                rc = OscapAtomicScan.scan_cve(OS, args.scan_target, leftover_args)
-            else:
-                parser.print_help()
-                sys.exit(2)
-
-        else:  # without atomic
-            print("Using native Docker API")
-
-            ODS = OscapDockerScan(args.scan_target, args.is_image, args.oscap_binary)
-            if args.action == "scan":
-                rc = OscapDockerScan.scan(ODS, leftover_args)
-            elif args.action == "scan_cve":
-                rc = OscapDockerScan.scan_cve(ODS, leftover_args)
-            else:
-                parser.print_help()
-                sys.exit(2)
+        ODS = OscapDockerScan(args.scan_target, args.is_image, args.oscap_binary)
+        if args.action == "scan":
+            rc = OscapDockerScan.scan(ODS, leftover_args)
+        elif args.action == "scan_cve":
+            rc = OscapDockerScan.scan_cve(ODS, leftover_args)
+        else:
+            parser.print_help()
+            sys.exit(2)
 
     except (ValueError, RuntimeError) as e:
         raise e
